requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
aio-pika==8.2.4
    # via wipac-keycloak-rest-services (setup.py)
aiormq==6.4.2
    # via aio-pika
cachetools==5.2.0
    # via
    #   google-auth
    #   wipac-keycloak-rest-services (setup.py)
    #   wipac-rest-tools
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==38.0.1
    # via pyjwt
dnspython==2.2.1
    # via pymongo
google-api-core==2.10.2
    # via google-api-python-client
google-api-python-client==2.65.0
    # via wipac-keycloak-rest-services (setup.py)
google-auth==2.13.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.7.0
    # via wipac-keycloak-rest-services (setup.py)
googleapis-common-protos==1.56.4
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   requests
    #   yarl
ldap3==2.9.1
    # via wipac-keycloak-rest-services (setup.py)
motor==3.1.1
    # via wipac-keycloak-rest-services (setup.py)
multidict==6.0.2
    # via yarl
oauthlib==3.2.2
    # via requests-oauthlib
pamqp==3.2.1
    # via aiormq
protobuf==4.21.8
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   ldap3
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.5.0
    # via
    #   wipac-keycloak-rest-services (setup.py)
    #   wipac-rest-tools
pymongo==4.3.2
    # via motor
pyparsing==3.0.9
    # via httplib2
requests==2.28.1
    # via
    #   google-api-core
    #   requests-futures
    #   requests-oauthlib
    #   wipac-dev-tools
    #   wipac-keycloak-rest-services (setup.py)
    #   wipac-rest-tools
requests-futures==1.0.0
    # via wipac-rest-tools
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
tornado==6.2
    # via wipac-rest-tools
types-cryptography==3.3.23.1
    # via pyjwt
typing-extensions==4.4.0
    # via wipac-dev-tools
unidecode==1.3.6
    # via wipac-keycloak-rest-services (setup.py)
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
wipac-dev-tools==1.6.0
    # via
    #   wipac-keycloak-rest-services (setup.py)
    #   wipac-rest-tools
wipac-rest-tools==1.3.9
    # via wipac-keycloak-rest-services (setup.py)
yarl==1.8.1
    # via
    #   aio-pika
    #   aiormq
